Decrypt packet and validate before collection.
The reason to decrypt at the server side (instead of application side) is because it is hard / impossible to separate these responsibilities. E.g. the JoinAccept message is encrypted with the AppKey and contains the DevAddr. This means that when the application side needs to encrypt this payload, assigning the DevAddr becomes the responsibility of the application, not the server.